Create a serverless WASM application
The application is pretty simple and validates a HS256 signed JWT
against a secret and rate-limits calls based on the client IP.
We could not use any library that binds against C lib since this is not
supported for wasm tagets, this also means that a lot of Rust crates for
JWT validation and or crypto were not usable because they require either
openssl bindings or use the ring crate.
A rate-limiting example was implemented to use the key-value store
provided by fastly's http_guest interface.
At the moment, the documentation is quite lacking, e.g. the user-agent
must not be overwritten otherwise some random 'Send error` is returned.
Also, deployment was kind of hard until we discovered terrctl.